removed redundant constructors in car

+++ tests/CarUnitTests.java
@@ -49,7 +49,7 @@ public class CarUnitTests {
     @Test
     public void isEmptySensor1(){
     	
-    	double []spaces=new double[10];
+    	int []spaces=new int[10];
     	for (int i =0; i<10;i++) 
             spaces[i] = 100;
         
@@ -59,7 +59,6 @@ public class CarUnitTests {
     	double[] test=pa.isEmpty();
     	//System.out.println(test[0]);
     	//System.out.println(test[1]);
-    	//System.out.println("sensor "+test[0]);
     	assertEquals(100.0,test[0],15);
     }
     
@@ -67,7 +66,7 @@ public class CarUnitTests {
     @Test
     public void isEmptySensor2(){
     	
-    	double []spaces=new double[10];
+    	int []spaces=new int[10];
     	for (int i =0; i<10;i++) 
             spaces[i] = 100;
         
@@ -100,7 +99,7 @@ public class CarUnitTests {
     public void MoveBackward(){
     
     	Car pa = new Car(100,false);
-    	pa.generateMap(new double[0]);
+    	pa.generateMap(new int[0]);
     	State state=pa.MoveBackward();
     	assertEquals(99,state.position);
     }
@@ -109,7 +108,7 @@ public class CarUnitTests {
     @Test
     public void MoveBackwardOutOfBounds(){
     	Car pa = new Car(0,false);
-    	pa.generateMap(new double[0]);
+    	pa.generateMap(new int[0]);
     	State state=pa.MoveBackward();
     	assertEquals(0,state.position);
     }
@@ -136,7 +135,7 @@ public class CarUnitTests {
     @Test
     public void oneSensorDisabledPark(){
     	
-    	double []spaces=new double[15];
+    	int []spaces=new int[15];
     	for (int i =0; i<10;i++) 
             spaces[i] = 100;
     	for (int i =10; i<15;i++) 
@@ -146,7 +145,7 @@ public class CarUnitTests {
     	pa.disableSensor(-1, 13);
     	pa.generateMap(spaces);
     	pa.Park();
-    	assertTrue(pa.WhereIs().isParked && pa.WhereIs().position==14.0);
+    	assertTrue(pa.WhereIs().isParked && pa.WhereIs().position==14);
 
 
     }
@@ -156,7 +155,7 @@ public class CarUnitTests {
     public void allSensorsDiabledPark(){
     	
     	
-    	double []spaces=new double[16];
+    	int []spaces=new int[16];
     	for (int i =0; i<10;i++) 
             spaces[i] = 100;
     	for (int i =10; i<15;i++) 
@@ -175,7 +174,7 @@ public class CarUnitTests {
     @Test
     public void Park(){
     	
-    	double []spaces={100,100,100,100,100,178,100,100,178,178,178,178,100,178,178,178,178,178};
+    	int []spaces={100,100,100,100,100,178,100,100,178,178,178,178,100,178,178,178,178,178};
     
     	Car pa = new Car(0,false);
     	pa.generateMap(spaces);
@@ -186,7 +185,7 @@ public class CarUnitTests {
     
     @Test
     public void ParkInFirstSpace(){
-    	double []spaces=new double[15];
+    	int []spaces=new int[15];
     	for (int i =0; i<5;i++) spaces[i] = 100;
     	for (int i =5; i<10;i++) spaces[i] = 178;
